% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensShiny.R
\docType{package}
\name{SensShiny}
\alias{SensShiny}
\alias{SensShiny-package}
\title{Shiny solutions for Sensory Analysis and Consumer Tests}
\description{
The SensShiny package contains wrappers functions for
    building web interfaces, implemented in shiny, to collect data
    from sensory and consumer experiments.
}
